Updated error and added description to db

+++ controller.py
@@ -4,6 +4,8 @@ from flask_restful import Resource
 from dbconfig import db
 from models import Task
 
+import error
+
 
 class Welcome(Resource):
     @staticmethod
@@ -16,13 +18,17 @@ class GetTasks(Resource):
     def get():
         try:
             tasks = Task.query.all()
+            if not tasks:
+                return error.NO_INPUT_400
             tasks_list = [
                 {
                     'task_name': task.task_name,
+                    'description': task.description,
                     'created_date': task.created_date.strftime('%Y-%m-%d %H:%M:%S') if task.created_date else None
                 }
                 for task in tasks
             ]
+
             return tasks_list
         except Exception as e:
             return str(e)
@@ -33,7 +39,11 @@ class AddTask(Resource):
     def post():
         try:
             data = request.json
-            task = Task(task_name=data.get('task_name'))
+            task = data.get('task_name')
+            description = data.get('description')
+            if not task or not description:
+                return error.DATA_NOT_FOUND
+            task = Task(task_name=task, description=description)
             db.session.add(task)
             db.session.commit()
             return GetTasks.get(), 201
@@ -48,7 +58,12 @@ class UpdateTask(Resource):
             data = request.json
             task_id = request.args.get('id')
             task = Task.query.filter_by(id=task_id).first()
-            task.task_name = data['task_name']
+            if not task:
+                return error.NOT_FOUND_404
+            if 'task_name' in data:
+                task.task_name = data['task_name']
+            if 'description' in data:
+                task.description = data['description']
             db.session.commit()
             return GetTasks.get()
         except Exception as e:
@@ -61,6 +76,8 @@ class DeleteTask(Resource):
         try:
             task_id = request.args.get('id')
             task = Task.query.filter_by(id=task_id).first()
+            if task is None:
+                return error.DOES_NOT_EXIST
             db.session.delete(task)
             db.session.commit()
             return GetTasks.get()
